requirements.txt
keyboard
pyserial
numpy
matplotlib
pandas
